+++ src/four_mat/four_vec/mod.rs
@@ -1,4 +1,5 @@
 use std::f64;
+use self::f64::NAN;
 use std::ops::Add;
 use std::ops::AddAssign;
 use std::ops::Sub;
@@ -7,20 +8,20 @@ use std::ops::Mul;
 use std::ops::Neg;
 use std::iter;
 use std::fmt;
-use std::str::FromStr;
-use std::num::ParseFloatError;
+use std::error;
 
 use crate::three_mat;
-use crate::utils;
 
 use three_mat::ThreeVec;
 
+use crate::utils;
 use utils::consts;
-use utils::Serializable;
+use utils::{Serializable, Deserializable};
 use utils::errors::CalcifyError;
 
 extern crate rmp;
 use rmp::encode::*;
+use rmp::decode::*;
 
 /// Variants of S space-time invariant
 #[derive(Debug, PartialEq)]
@@ -101,6 +102,31 @@ impl FourVec {
             m1,
             m2,
             m3,
+        }
+    }
+
+    /// Returns a new FourVec from a slice
+    ///
+    /// # Arguments
+    ///
+    /// * `slice` - &[f64]
+    ///
+    /// # Example
+    /// ```
+    /// use calcify::FourVec;
+    /// let vec4: FourVec = FourVec::from(&[1.0,1.0,1.0,1.0]);
+    /// ```
+    ///
+    /// # Panics
+    ///
+    /// * `slice` length < 4
+    pub fn from(slice: &[f64]) -> FourVec {
+
+        FourVec {
+            m0: slice[0],
+            m1: slice[1],
+            m2: slice[2],
+            m3: slice[3],
         }
     }
 
@@ -245,9 +271,6 @@ impl Serializable for FourVec {
         format!("{{\"m0\":{},\"m1\":{},\"m2\":{},\"m3\":{}}}",
             self.m0(), self.m1(), self.m2(), self.m3())
     }
-    fn to_jsonc(&self) -> String {
-        format!("[{},{},{},{}]", self.m0(), self.m1(), self.m2(), self.m3())
-    }
     fn to_msg(&self) -> Result<Vec<u8>,ValueWriteError> {
         let mut buf = Vec::with_capacity(5);
         write_array_len(&mut buf, 4)?;
@@ -259,14 +282,13 @@ impl Serializable for FourVec {
     }
 }
 
-impl FromStr for FourVec {
-    type Err = ParseFloatError;
+impl Deserializable for FourVec {
 
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut m0: f64 = std::f64::NAN;
-        let mut m1: f64 = std::f64::NAN;
-        let mut m2: f64 = std::f64::NAN;
-        let mut m3: f64 = std::f64::NAN;
+    fn from_json(s: &str) -> Result<Self, Box<dyn error::Error>> {
+        let mut m0: f64 = NAN;
+        let mut m1: f64 = NAN;
+        let mut m2: f64 = NAN;
+        let mut m3: f64 = NAN;
         for dim in s.trim_matches(|p| p == '{' || p == '}' ).split(',') {
             let n_v: Vec<&str> = dim.split(':').collect();
             match n_v[0] {
@@ -274,12 +296,25 @@ impl FromStr for FourVec {
                 "\"m1\"" => m1 = n_v[1].parse::<f64>()?,
                 "\"m2\"" => m2 = n_v[1].parse::<f64>()?,
                 "\"m3\"" => m3 = n_v[1].parse::<f64>()?,
-                x => panic!("Unexpected invalid token {:?}", x),
+                _ => return Err(Box::new(CalcifyError::ParseError)),
             }
         }
         Ok(FourVec{m0,m1,m2,m3})
     }
+
+    fn from_msg(mut bytes: &[u8]) -> Result<(Self,&[u8]), Box<dyn error::Error>> {
+        if let Ok(4) = read_array_len(&mut bytes){
+            let mut x: [f64;4] = [NAN;4];
+            for i in 0..4 {
+                x[i] = read_f64(&mut bytes)?;
+            }
+            Ok((FourVec::from(&x),bytes))
+        } else {
+            Err(Box::new(CalcifyError::ParseError))
+        }
+    }
 }
+
 
 
 impl Add for FourVec {
@@ -427,6 +462,14 @@ mod tests {
     fn test_parse() {
         let xx = FourVec::new(5.0,2.0,2.0,2.0);
         let pp = xx.to_json();
-        assert_eq!(FourVec::from_str(&pp).unwrap(),xx);
+        assert_eq!(FourVec::from_json(&pp).unwrap(),xx);
+    }
+
+    #[test]
+    fn test_msg_parse() {
+        let xx = FourVec::new(5.0,2.0,2.0,2.0);
+        let pp = xx.to_msg().unwrap();
+        let (oo,_) = FourVec::from_msg(&pp).unwrap();
+        assert_eq!(oo,xx);
     }
 }