+++ src/four_mat/mod.rs
@@ -1,3 +1,5 @@
+use std::f64::NAN;
+
 use std::ops::Add;
 use std::ops::AddAssign;
 use std::ops::Sub;
@@ -5,8 +7,7 @@ use std::ops::SubAssign;
 use std::ops::Mul;
 use std::ops::Neg;
 use std::fmt;
-use std::str::FromStr;
-use std::num::ParseFloatError;
+use std::error;
 
 mod four_vec;
 
@@ -20,11 +21,12 @@ use crate::utils;
 
 use three_mat::ThreeVec;
 
-use utils::Serializable;
+use utils::{Serializable, Deserializable};
 use utils::errors::CalcifyError;
 
 extern crate rmp;
 use rmp::encode::*;
+use rmp::decode::*;
 
 /// Four Matrix
 #[derive(Debug, PartialEq, Copy, Clone)]
@@ -63,6 +65,25 @@ impl FourMat {
             n1,
             n2,
             n3,
+        }
+    }
+
+    /// Returns a new FourVec from a slice
+    ///
+    /// # Arguments
+    ///
+    /// * `slice` - &[FourVec]
+    ///
+    /// # Panics
+    ///
+    /// * `slice` length < 4
+    pub fn from(slice: &[FourVec]) -> FourMat {
+
+        FourMat {
+            n0: slice[0],
+            n1: slice[1],
+            n2: slice[2],
+            n3: slice[3],
         }
     }
 
@@ -274,14 +295,7 @@ impl Serializable for FourMat {
             self.n3().to_json()
         )
     }
-    fn to_jsonc(&self) -> String {
-        format!("[{},{},{},{}]",
-            self.n0().to_jsonc(),
-            self.n1().to_jsonc(),
-            self.n2().to_jsonc(),
-            self.n3().to_jsonc()
-        )
-    }
+
     fn to_msg(&self) -> Result<Vec<u8>,ValueWriteError> {
         let mut buf = Vec::new();
         write_array_len(&mut buf, 4)?;
@@ -294,25 +308,38 @@ impl Serializable for FourMat {
 
 }
 
-impl FromStr for FourMat {
-    type Err = ParseFloatError;
+impl Deserializable for FourMat {
 
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut n0: FourVec = FourVec::new(std::f64::NAN,std::f64::NAN,std::f64::NAN,std::f64::NAN);
-        let mut n1: FourVec = FourVec::new(std::f64::NAN,std::f64::NAN,std::f64::NAN,std::f64::NAN);
-        let mut n2: FourVec = FourVec::new(std::f64::NAN,std::f64::NAN,std::f64::NAN,std::f64::NAN);
-        let mut n3: FourVec = FourVec::new(std::f64::NAN,std::f64::NAN,std::f64::NAN,std::f64::NAN);
+    fn from_json(s: &str) -> Result<Self, Box<dyn error::Error>> {
+        let mut n0: FourVec = FourVec::new(NAN,NAN,NAN,NAN);
+        let mut n1: FourVec = FourVec::new(NAN,NAN,NAN,NAN);
+        let mut n2: FourVec = FourVec::new(NAN,NAN,NAN,NAN);
+        let mut n3: FourVec = FourVec::new(NAN,NAN,NAN,NAN);
         for dim in s.replace("}}","|}").replace("},","}|").replace(":{",":!{").trim_matches(|p| p == '{' || p == '}' ).split_terminator('|') {
             let n_v: Vec<&str> = dim.split(":!").collect();
             match n_v[0] {
-                "\"n0\"" => n0 = FourVec::from_str(n_v[1])?,
-                "\"n1\"" => n1 = FourVec::from_str(n_v[1])?,
-                "\"n2\"" => n2 = FourVec::from_str(n_v[1])?,
-                "\"n3\"" => n3 = FourVec::from_str(n_v[1])?,
-                x => panic!("Unexpected invalid token {:?}", x),
+                "\"n0\"" => n0 = FourVec::from_json(n_v[1])?,
+                "\"n1\"" => n1 = FourVec::from_json(n_v[1])?,
+                "\"n2\"" => n2 = FourVec::from_json(n_v[1])?,
+                "\"n3\"" => n3 = FourVec::from_json(n_v[1])?,
+                _ => return Err(Box::new(CalcifyError::ParseError)),
             }
         }
         Ok(FourMat{n0,n1,n2,n3})
+    }
+
+    fn from_msg(mut bytes: &[u8]) -> Result<(Self,&[u8]), Box<dyn error::Error>> {
+        if let Ok(4) = read_array_len(&mut bytes){
+            let mut x: [FourVec;4] = [FourVec::new(NAN,NAN,NAN,NAN);4];
+            for i in 0..4 {
+                let (vec,rest) = FourVec::from_msg(&mut bytes)?;
+                x[i] = vec;
+                bytes = rest;
+            }
+            Ok((FourMat::from(&x),bytes))
+        } else {
+            Err(Box::new(CalcifyError::ParseError))
+        }
     }
 }
 
@@ -516,6 +543,17 @@ mod tests {
                                     FourVec::new(1.0,1.0,1.0,1.0),
                                     FourVec::new(1.0,1.0,1.0,1.0));
         let pp = xx.to_json();
-        assert_eq!(FourMat::from_str(&pp).unwrap(),xx);
+        assert_eq!(FourMat::from_json(&pp).unwrap(),xx);
+    }
+
+    #[test]
+    fn test_msg_parse() {
+        let xx = FourMat::new(FourVec::new(1.0,1.0,1.0,1.0),
+                                    FourVec::new(1.0,2.0,1.0,1.0),
+                                    FourVec::new(1.0,1.0,3.0,1.0),
+                                    FourVec::new(1.0,1.0,1.0,4.0));
+        let pp = xx.to_msg().unwrap();
+        let (oo,_) = FourMat::from_msg(&pp).unwrap();
+        assert_eq!(oo,xx);
     }
 }