+++ src/utils/mod.rs
@@ -3,6 +3,7 @@ pub mod consts;
 
 pub mod serializable;
 pub use serializable::Serializable;
+pub use serializable::Deserializable;
 
 /// Errors  module
 pub mod errors;